# app_express/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pembukuan TBS API",
    "version": "1.0.0",
    "description": "REST API untuk aplikasi Pembukuan TBS (manajemen slip timbangan kelapa sawit).\n\nAutentikasi via header `x-api-key`."
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1" }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "Bon": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "ticket_number": { "type": "string" },
          "bon_date": { "type": "string", "format": "date" },
          "plate_number": { "type": "string" },
          "driver_name": { "type": "string" },
          "relation_name": { "type": "string" },
          "relation_agent_id": { "type": "string" },
          "factory_id": { "type": "string" },
          "factory_spsi_type_id": { "type": "string" },
          "spsi_type_name": { "type": "string" },
          "spsi_calculation_mode": { "type": "string", "enum": ["PER_KG", "FIX"] },
          "spsi_rate": { "type": "number" },
          "spsi_amount": { "type": "number" },
          "fruit_origin": { "type": "string" },
          "netto_1": { "type": "number" },
          "netto_2": { "type": "number" },
          "price": { "type": "number" },
          "dp": { "type": "number" },
          "biaya_bongkar": { "type": "number" },
          "bp_colt": { "type": "number" },
          "pph": { "type": "number" },
          "uang_minum": { "type": "number" },
          "notes": { "type": "string", "description": "Catatan" },
          "total": { "type": "number" },
          "image_url": { "type": "string" },
          "status": { "type": "string", "enum": ["BELUM_DIBAYAR", "TERTAGIH", "LUNAS"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "bon_deductions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "bon_id": { "type": "string" },
                "label": { "type": "string" },
                "amount": { "type": "number" }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
    "/bons": {
      "get": {
        "summary": "Daftar bon",
        "tags": ["Bon"],
        "parameters": [
          { "in": "query", "name": "start", "schema": { "type": "string", "format": "date" }, "description": "Tanggal awal (YYYY-MM-DD)" },
          { "in": "query", "name": "end", "schema": { "type": "string", "format": "date" }, "description": "Tanggal akhir (YYYY-MM-DD)" },
          { "in": "query", "name": "q", "schema": { "type": "string" }, "description": "Cari driver, plat, atau relasi" },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["BELUM_DIBAYAR", "TERTAGIH", "LUNAS"] }, "description": "Filter status" }
        ],
        "responses": { "200": { "description": "Daftar bon", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Bon" } } } } } }
      },
      "post": {
        "summary": "Buat bon baru",
        "tags": ["Bon"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary", "description": "Foto bon" },
                  "ticket_number": { "type": "string" },
                  "bon_date": { "type": "string", "format": "date" },
                  "plate_number": { "type": "string" },
                  "driver_name": { "type": "string" },
                  "relation_name": { "type": "string" },
                  "relation_agent_id": { "type": "string" },
                  "factory_id": { "type": "string" },
                  "factory_spsi_type_id": { "type": "string" },
                  "netto_1": { "type": "number" },
                  "netto_2": { "type": "number" },
                  "price": { "type": "number" },
                  "dp": { "type": "number" },
                  "biaya_bongkar": { "type": "number" },
                  "bp_colt": { "type": "number" },
                  "uang_minum": { "type": "number" },
                  "chat_id": { "type": "string", "description": "ID chat asal (dari Telegram/WA)" },
                  "message_id": { "type": "string", "description": "ID message asal (dari Telegram/WA)" },
                  "notes": { "type": "string", "description": "Catatan" },
                  "deduction_label": { "type": "array", "items": { "type": "string" }, "description": "Label potongan dinamis" },
                  "deduction_amount": { "type": "array", "items": { "type": "number" }, "description": "Jumlah potongan dinamis" }
                },
                "required": ["bon_date", "plate_number", "relation_agent_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Bon created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Bon" } } } }
        }
      }
    },
...
    "/bons/{id}": {
      "get": {
        "summary": "Detail bon + relasi",
        "tags": ["Bon"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Detail bon", "content": { "application/json": { "schema": { "type": "object", "properties": { "bon": { "$ref": "#/components/schemas/Bon" }, "related": { "type": "object" } } } } } }
        }
      },
      "patch": {
        "summary": "Update bon",
        "tags": ["Bon"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary" },
                  "plate_number": { "type": "string" },
                  "driver_name": { "type": "string" },
                  "netto_1": { "type": "number" },
                  "netto_2": { "type": "number" },
                  "price": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Bon updated" } }
      },
      "delete": {
        "summary": "Hapus bon",
        "tags": ["Bon"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Bon deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "deleted": { "type": "boolean" } } } } } } }
      }
    },
...
  }
}